Moved CheckoutError and CheckoutCancel out of page
